backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.15
psycopg2-binary>=2.9.6
pydantic>=1.10.7
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0
fastapi-mail>=1.2.8
alembic>=1.10.3
pydantic-settings>=2.0.2
httpx>=0.24.0
authlib>=1.2.0
itsdangerous>=2.1.2
PyJWT>=2.8.0